web_project/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==2.0.0
flask-migrate==4.0.4
ultralytics==8.0.120
opencv-python==4.7.0.72
pillow==10.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-socketio==5.3.5
requests==2.31.0
flask-mail==0.9.1